backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
pydantic-settings==2.5.2
telethon==1.36.0
rich==13.9.2
pwinput==1.0.3
python-multipart==0.0.12
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
